+++ src/test/kotlin/io/kstuff/test/GenericTestsTest.kt
@@ -1,5 +1,5 @@
 /*
- * @(#) TestFunctionsTest.kt
+ * @(#) GenericTestsTest.kt
  *
  * should-test  Kotlin testing functions
  * Copyright (c) 2024 Peter Wall
@@ -28,9 +28,8 @@ package io.kstuff.test
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
-import kotlin.test.assertFalse
 
-class TestFunctionsTest {
+class GenericTestsTest {
 
     @Test fun `should perform shouldBe`() {
         value1 shouldBe 12345
@@ -56,13 +55,6 @@ class TestFunctionsTest {
         }
         assertFailsWith<AssertionError> { value6 shouldBe null }.let {
             assertEquals("Value should be null, was 12345", it.message)
-        }
-    }
-
-    @Test fun `should handle errors involving long strings differently`() {
-        assertFailsWith<AssertionError> { longString shouldBe anotherLongString }.let {
-            val message = it.message
-            assertFalse(message != null && message.startsWith("Value should be "))
         }
     }
 
@@ -107,137 +99,27 @@ class TestFunctionsTest {
         }
     }
 
-    @Test fun `should perform shouldContain`() {
-        list1 shouldContain "alpha"
-        list1 shouldContain "gamma"
+    @Test fun `should perform shouldBeOneOf`() {
+        value1 shouldBeOneOf listOf(1234, 12345, 123456)
+        value2 shouldBeOneOf listOf("Test string", "Dummy string")
     }
 
     @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldContain fails`() {
-        assertFailsWith<AssertionError> { list1 shouldContain "delta" }.let {
-            assertEquals("Collection should contain \"delta\"", it.message)
+    @Test fun `should throw AssertionError when shouldBeOneOf fails`() {
+        assertFailsWith<AssertionError> { value1 shouldBeOneOf listOf(1, 2, 3) }.let {
+            assertEquals("Value should be one of collection, was 12345", it.message)
         }
     }
 
-    @Test fun `should perform shouldNotContain`() {
-        list1 shouldNotContain "delta"
+    @Test fun `should perform shouldNotBeOneOf`() {
+        value1 shouldNotBeOneOf listOf(1, 2, 3)
+        value2 shouldNotBeOneOf listOf("alpha", "beta", "gamma")
     }
 
     @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldNotContain fails`() {
-        assertFailsWith<AssertionError> { list1 shouldNotContain "alpha" }.let {
-            assertEquals("Collection should not contain \"alpha\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldContainKey`() {
-        map1 shouldContainKey "first"
-        map1 shouldContainKey "second"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldContainKey fails`() {
-        assertFailsWith<AssertionError> { map1 shouldContainKey "alpha" }.let {
-            assertEquals("Map should contain key \"alpha\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldNotContainKey`() {
-        map1 shouldNotContainKey "alpha"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldNotContainKey fails`() {
-        assertFailsWith<AssertionError> { map1 shouldNotContainKey "first" }.let {
-            assertEquals("Map should not contain key \"first\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldContain on string`() {
-        value2 shouldContain "st"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldContain on string fails`() {
-        assertFailsWith<AssertionError> { value2 shouldContain "abc" }.let {
-            assertEquals("String should contain \"abc\", was \"Test string\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldNotContain on string`() {
-        value2 shouldNotContain "stop"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldNotContain on string fails`() {
-        assertFailsWith<AssertionError> { value2 shouldNotContain "st" }.let {
-            assertEquals("String should not contain \"st\", was \"Test string\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldStartWith`() {
-        value2 shouldStartWith "Test"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldStartWith fails`() {
-        assertFailsWith<AssertionError> { value2 shouldStartWith "X" }.let {
-            assertEquals("String should start with \"X\", was \"Test string\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldNotStartWith`() {
-        value2 shouldNotStartWith "X"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldNotStartWith fails`() {
-        assertFailsWith<AssertionError> { value2 shouldNotStartWith "T" }.let {
-            assertEquals("String should not start with \"T\", was \"Test string\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldEndWith`() {
-        value2 shouldEndWith "string"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldEndWith fails`() {
-        assertFailsWith<AssertionError> { value2 shouldEndWith "X" }.let {
-            assertEquals("String should end with \"X\", was \"Test string\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldNotEndWith`() {
-        value2 shouldNotEndWith "X"
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldNotEndWith fails`() {
-        assertFailsWith<AssertionError> { value2 shouldNotEndWith "string" }.let {
-            assertEquals("String should not end with \"string\", was \"Test string\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldMatch`() {
-        value2 shouldMatch Regex("^T")
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldMatch fails`() {
-        assertFailsWith<AssertionError> { value2 shouldMatch Regex("^X") }.let {
-            assertEquals("String should match regex ^X, was \"Test string\"", it.message)
-        }
-    }
-
-    @Test fun `should perform shouldNotMatch`() {
-        value2 shouldNotMatch Regex("^X")
-    }
-
-    @Suppress("ComplexRedundantLet")
-    @Test fun `should throw AssertionError when shouldNotMatch fails`() {
-        assertFailsWith<AssertionError> { value2 shouldNotMatch Regex("^T") }.let {
-            assertEquals("String should not match regex ^T, was \"Test string\"", it.message)
+    @Test fun `should throw AssertionError when shouldNotBeOneOf fails`() {
+        assertFailsWith<AssertionError> { value1 shouldNotBeOneOf listOf(12345, 12346) }.let {
+            assertEquals("Value should not be one of collection, was 12345", it.message)
         }
     }
 
@@ -275,6 +157,18 @@ class TestFunctionsTest {
         assertFailsWith<AssertionError> { shouldThrow<NumberFormatException> { "123".toInt() } }
     }
 
+    @Test fun `should perform shouldThrow with message`() {
+        shouldThrow<RuntimeException>("Something went wrong") { throw RuntimeException("Something went wrong") }
+    }
+
+    @Test fun `should throw AssertionError when shouldThrow with message fails`() {
+        assertFailsWith<AssertionError> {
+            shouldThrow<RuntimeException>("Something went wrong") { throw RuntimeException("Bad message") }
+        }.let {
+            assertEquals("Message incorrect, was \"Bad message\"", it.message)
+        }
+    }
+
     @Test fun `should perform shouldBeNonNull`() {
         value6.shouldBeNonNull() shouldBe 12345
     }
@@ -309,19 +203,6 @@ class TestFunctionsTest {
         val value5: String? = null
         @Suppress("RedundantNullableReturnType")
         val value6: Int? = 12345
-        val list1 = listOf("alpha", "beta", "gamma")
-        val map1 = mapOf("first" to "alpha", "second" to "beta", "third" to "gamma")
-
-        const val longString =
-"""Our revels now are ended. These our actors,
-As I foretold you, were all spirits and
-Are melted into air, into thin air:
-"""
-        const val anotherLongString =
-"""We are such stuff
-As dreams are made on, and our little life
-Is rounded with a sleep.
-"""
 
     }