+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## [2.0] - 2024-11-26
+## [2.1] - 2024-11-26
 ### Added
+- `CollectionTests`, `StringTests`: split out from `TestFunctions`
+### Changed
+- `TestFunctions`: renamed to `GenericTests`
+- `GenericTests`: added `shouldBeOneOf` and `shouldNotBeOneOf`
+- `GenericTests`: added `shouldThrow` with message
+
+## [2.0] - 2024-11-26
+### Changed
 - `TestFunctions`: extensive enhancements
 
 ## [1.0] - 2024-11-24